Rename test/util.ts to test/utils.ts
To match src/utils.ts
